intial adaptation, next step test
