+++ src/main/java/com/olivar/springbootmongodbdocker/dao/ManufacturerRepo.java
@@ -1,6 +1,6 @@
 package com.olivar.springbootmongodbdocker.dao;
 
-import com.olivar.springbootmongodbdocker.model.Manufacturer;
+import com.olivar.springbootmongodbdocker.domain.model.Manufacturer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface ManufacturerRepo extends MongoRepository<Manufacturer,String> {
