+++ src/main/java/com/olivar/springbootmongodbdocker/infrastructure/inboud/rest/core/dto/ManufacturerResponse.java
@@ -1,0 +1,9 @@
+package com.olivar.springbootmongodbdocker.infrastructure.inboud.rest.core.dto;
+
+import lombok.Data;
+
+@Data
+public class ManufacturerResponse {
+    private String name;
+    private String code;
+}
